fetch new update from hp13 workstation for better precision.

+++ Trunk/Win32/wav44to16/Coeffs.h
@@ -5,7 +5,7 @@
 #define TARGET_SAMPLE_RATE  (16000)  //Hz
 #define SOURCE_SAMPLE_RATE  (44100)  //Hz
 
-const float FirCoeffs[] =
+const double FirCoeffs[] =
 {
 #include "FirCoeff44To16.dat"
 };
